requirements.txt
fastapi
uvicorn
transformers
torch
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
python-dotenv
mysql-connector-python